refactor: TrialMember 엔티티 JPA 맵핑 수정 등
- Trial 엔티티에 새 필드 추가
- TrialMember 엔티티에 새 필드 추가
- Member 테이블 스키마 id 타입 INT -> BIGINT로 변경

trial_member 테이블에 idx_member_id 인덱스 추가

+++ solomon/src/main/resources/db/V1.02__create_table_trial.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS trial (
     stage VARCHAR(255) NOT NULL,
 
     last_message TEXT,
-    last_message_seq INT,
+    last_message_seq BIGINT,
 
     created_at DATETIME NOT NULL,
     updated_at DATETIME NOT NULL,
